move setup_utils_config to configuration module
