+++ freqtrade/configuration/__init__.py
@@ -1,5 +1,6 @@
 # flake8: noqa: F401
 
+from freqtrade.configuration.config_setup import setup_utils_configuration
 from freqtrade.configuration.check_exchange import check_exchange, remove_credentials
 from freqtrade.configuration.timerange import TimeRange
 from freqtrade.configuration.configuration import Configuration
