+++ freqtrade/configuration/config_setup.py
@@ -1,8 +1,9 @@
 import logging
 from typing import Any, Dict
 
-from freqtrade.configuration import (Configuration, remove_credentials,
-                                     validate_config_consistency)
+from .config_validation import validate_config_consistency
+from .configuration import Configuration
+from .check_exchange import remove_credentials
 from freqtrade.state import RunMode
 
 logger = logging.getLogger(__name__)
